refactor(lsp): split message dispatch module
