+++ crates/tsz-cli/src/bin/tsz_lsp.rs
@@ -69,6 +69,7 @@ use tsz::lsp::{
     CompletionItemData, CompletionItemKind, FormattingOptions, Position, Project, Range,
 };
 
+mod tsz_lsp_dispatch;
 mod tsz_lsp_request_handlers;
 
 /// TSZ Language Server
@@ -274,281 +275,6 @@ impl LspServer {
                 "message": message,
             }),
         });
-    }
-
-    // ─── Message dispatch ───────────────────────────────────────────────
-
-    fn handle_message(&mut self, msg: JsonRpcMessage) -> Option<JsonRpcResponse> {
-        let method = msg.method.as_deref();
-        let id = msg.id.clone();
-
-        // Handle cancellation notification
-        if method == Some("$/cancelRequest") {
-            self.handle_cancel_request(msg.params);
-            return None;
-        }
-
-        // Check if this request was already cancelled
-        if self.is_cancelled(&id)
-            && let Some(id_val) = id
-        {
-            let id_str = match &id_val {
-                Value::Number(n) => n.to_string(),
-                Value::String(s) => s.clone(),
-                _ => String::new(),
-            };
-            self.cancelled_requests.remove(&id_str);
-            return Some(self.error_response(
-                Some(id_val),
-                -32800,
-                "Request cancelled".to_string(),
-            ));
-        }
-
-        match method {
-            Some("initialize") => {
-                let result = self.handle_initialize(msg.params.as_ref());
-                Some(self.success_response(id, result))
-            }
-            Some("initialized") => {
-                self.initialized = true;
-                self.handle_initialized();
-                None
-            }
-            Some("shutdown") => {
-                self.shutdown_requested = true;
-                Some(self.success_response(id, Value::Null))
-            }
-            Some("exit") => {
-                std::process::exit(i32::from(!self.shutdown_requested));
-            }
-
-            // ── Document lifecycle ──────────────────────────────────────
-            Some("textDocument/didOpen") => {
-                self.handle_did_open(msg.params);
-                None
-            }
-            Some("textDocument/didChange") => {
-                self.handle_did_change(msg.params);
-                None
-            }
-            Some("textDocument/didClose") => {
-                self.handle_did_close(msg.params);
-                None
-            }
-            Some("textDocument/didSave") => {
-                self.handle_did_save(msg.params);
-                None
-            }
-
-            // ── Language features ───────────────────────────────────────
-            Some("textDocument/hover") => {
-                let r = self.handle_hover(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/completion") => {
-                let r = self.handle_completion(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("completionItem/resolve") => {
-                let r = self.handle_completion_resolve(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/definition") | Some("textDocument/declaration") => {
-                let r = self.handle_definition(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/typeDefinition") => {
-                let r = self.handle_type_definition(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/references") => {
-                let r = self.handle_references(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/implementation") => {
-                let r = self.handle_implementation(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/documentSymbol") => {
-                let r = self.handle_document_symbol(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/formatting") => {
-                let r = self.handle_formatting(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/rename") => {
-                let r = self.handle_rename(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/prepareRename") => {
-                let r = self.handle_prepare_rename(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/codeAction") => {
-                let r = self.handle_code_action(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/codeLens") => {
-                let r = self.handle_code_lens(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("codeLens/resolve") => {
-                let r = self.handle_code_lens_resolve(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/selectionRange") => {
-                let r = self.handle_selection_range(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/foldingRange") => {
-                let r = self.handle_folding_range(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/signatureHelp") => {
-                let r = self.handle_signature_help(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/semanticTokens/full") => {
-                let r = self.handle_semantic_tokens_full(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/semanticTokens/range") => {
-                let r = self.handle_semantic_tokens_range(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/documentHighlight") => {
-                let r = self.handle_document_highlight(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/inlayHint") => {
-                let r = self.handle_inlay_hint(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("inlayHint/resolve") => {
-                let r = self.handle_inlay_hint_resolve(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/documentColor") => {
-                let r = self.handle_document_color(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/colorPresentation") => {
-                let r = self.handle_color_presentation(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/documentLink") => {
-                let r = self.handle_document_link(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/linkedEditingRange") => {
-                let r = self.handle_linked_editing_range(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/prepareCallHierarchy") => {
-                let r = self.handle_prepare_call_hierarchy(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("callHierarchy/incomingCalls") => {
-                let r = self.handle_incoming_calls(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("callHierarchy/outgoingCalls") => {
-                let r = self.handle_outgoing_calls(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("textDocument/prepareTypeHierarchy") => {
-                let r = self.handle_prepare_type_hierarchy(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("typeHierarchy/supertypes") => {
-                let r = self.handle_supertypes(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("typeHierarchy/subtypes") => {
-                let r = self.handle_subtypes(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("workspace/symbol") => {
-                let r = self.handle_workspace_symbol(msg.params);
-                Some(self.make_response(id, r))
-            }
-
-            // ── Range formatting ──────────────────────────────────────
-            Some("textDocument/rangeFormatting") => {
-                let r = self.handle_range_formatting(msg.params);
-                Some(self.make_response(id, r))
-            }
-
-            // ── On-type formatting ────────────────────────────────────
-            Some("textDocument/onTypeFormatting") => {
-                let r = self.handle_on_type_formatting(msg.params);
-                Some(self.make_response(id, r))
-            }
-
-            // ── Workspace notifications ───────────────────────────────
-            Some("workspace/didChangeConfiguration") => {
-                self.handle_did_change_configuration(msg.params);
-                None
-            }
-            Some("workspace/didChangeWatchedFiles") => {
-                self.handle_did_change_watched_files(msg.params);
-                None
-            }
-            Some("workspace/didChangeWorkspaceFolders") => {
-                self.handle_did_change_workspace_folders(msg.params);
-                None
-            }
-
-            // ── Execute command ────────────────────────────────────────
-            Some("workspace/executeCommand") => {
-                let r = self.handle_execute_command(msg.params);
-                Some(self.make_response(id, r))
-            }
-
-            // ── Diagnostic pull model (LSP 3.17) ─────────────────
-            Some("textDocument/diagnostic") => {
-                let r = self.handle_document_diagnostic(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("workspace/diagnostic") => {
-                let r = self.handle_workspace_diagnostic(msg.params);
-                Some(self.make_response(id, r))
-            }
-
-            // ── File operations ─────────────────────────────────────
-            Some("workspace/willRenameFiles") => {
-                let r = self.handle_will_rename_files(msg.params);
-                Some(self.make_response(id, r))
-            }
-            Some("workspace/didRenameFiles") => {
-                self.handle_did_rename_files(msg.params);
-                None
-            }
-            Some("workspace/willCreateFiles") => {
-                // Acknowledge but no edits needed for file creation
-                Some(self.success_response(id, Value::Null))
-            }
-            Some("workspace/didCreateFiles") => {
-                self.handle_did_create_files(msg.params);
-                None
-            }
-            Some("workspace/willDeleteFiles") => {
-                // Acknowledge but no edits needed for file deletion
-                Some(self.success_response(id, Value::Null))
-            }
-            Some("workspace/didDeleteFiles") => {
-                self.handle_did_delete_files(msg.params);
-                None
-            }
-
-            // Unknown request → method not found
-            Some(method) if id.is_some() => {
-                Some(self.error_response(id, -32601, format!("Method not found: {method}")))
-            }
-            _ => None,
-        }
     }
 
     // ─── Response helpers ───────────────────────────────────────────────
